OutputControl:Files - Version Independednt INJECTION (Needs testing on dif v)
TODO 1:

Currently:
STOP BUTTON - Kills all EnergyPlus processes before terminating the worker processes.

Problem:
The user MAY have multiple EnergyPlus instances running simultaneously outside of this tool!

Solution:
So Leveraging EnergyPlus subprocess PID is crcial! And it has already been implemented in the past.

TODO 2:
Upon closing - kill eP_P.exe zomby processes.
Need to rebuild the exe.

TODO 3:
make the entire project together with the dependencies to be a single executable file, So it does not rquire an installation, and be all wrapped into a single file

TODO 4:
The checkboxesin the GUI should also be updated based on the SCHEMA.
So the schema-reading algorithm can be triggered once the user selected "Custom"